\name{CPDARMABC-package}
\alias{CPDARMABC-package}
\alias{CPDARMABC}
\docType{package}
\title{
\packageTitle{CPDARMABC}
}
\description{
\packageDescription{CPDARMABC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CPDARMABC}
\packageIndices{CPDARMABC}
  abc
}
\author{
\packageAuthor{CPDARMABC}

Maintainer: \packageMaintainer{CPDARMABC}
}
\references{
  abc
}
\keyword{ package }
\seealso{
   abc
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
