\name{distLoglikRatio}
\alias{distLoglikRatio}
\title{abc}
\usage{
distLoglikRatio(n, t, distPars0, lowerLambda, upperLambda, breakPoint, BetaFlg, W, Beta, crit, max_p, max_d, max_q, tol, maxIter, innovDist, nsim, maxErr)
}
\arguments{
  \item{n}{ abc }
  \item{t}{ abc }
  \item{distPars0}{ abc }
  \item{lowerLambda}{ abc }
  \item{upperLambda}{ abc }
  \item{breakPoint}{ abc }
  \item{BetaFlg}{ abc }
  \item{W}{ abc }
  \item{Beta}{ abc }
  \item{crit}{ abc }
  \item{max_p}{ abc }
  \item{max_d}{ abc }
  \item{max_q}{ abc }
  \item{tol}{ abc }
  \item{maxIter}{ abc }
  \item{innovDist}{ abc }
  \item{nsim}{ abc }
  \item{maxErr}{ abc }
}
\description{
	abc
}
\examples{
	\dontrun{
		# abc
	}
}