\name{simARIMAMat}
\alias{simARIMAMat}
\title{abc}
\usage{
simARIMAMat(mu, sigma2, innovDist, OmegaMat, BetaFlg, W, Beta)
}
\arguments{
  \item{mu}{ abc }
  \item{sigma2}{ abc }
  \item{innovDist}{ abc }
  \item{OmegaMat}{ abc }
  \item{BetaFlg}{ abc }
  \item{W}{ abc }
  \item{Beta}{ abc }
}
\description{
	abc
}
\examples{
	\dontrun{
		# abc
	}
}