\name{simARIMARec}
\alias{simARIMARec}
\title{abc}
\usage{
simARIMARec(model, n, mu, sigma2, innovDist, BetaFlg, W, Beta)
}
\arguments{
  \item{model}{ abc }
  \item{n}{ abc }
  \item{mu}{ abc }
  \item{sigma2}{ abc }
  \item{innovDist}{ abc }
  \item{BetaFlg}{ abc }
  \item{W}{ abc }
  \item{Beta}{ abc }
}
\description{
	abc
}
\examples{
	\dontrun{
		# abc
	}
}