\name{simCPDARIMABC}
\alias{simCPDARIMABC}
\title{abc}
\usage{
simCPDARIMABC(n, t, alpha, IC, model01, model02, GLRSApprox, lowerLambda, upperLambda, breakPoint, crit, max_p, max_d, max_q, tol, maxIter, innovDist, nsim1, nsim2, maxErr)
}
\arguments{
  \item{n}{ abc }
  \item{t}{ abc }
  \item{alpha}{ abc }
  \item{IC}{ abc }
  \item{model01}{ abc }
  \item{model02}{ abc }
  \item{GLRSApprox}{ abc }
  \item{lowerLambda}{ abc }
  \item{upperLambda}{ abc }
  \item{breakPoint}{ abc }
  \item{BetaFlg}{ abc }
  \item{W}{ abc }
  \item{Beta}{ abc }
  \item{crit}{ abc }
  \item{max_p}{ abc }
  \item{max_d}{ abc }
  \item{max_q}{ abc }
  \item{tol}{ abc }
  \item{maxIter}{ abc }
  \item{innovDist}{ abc }
  \item{nsim1}{ abc }
  \item{nsim2}{ abc }
  \item{maxErr}{ abc }
}
\description{
	abc
}
\examples{
	\dontrun{
		# abc
	}
}