\name{SldWin}
\alias{SldWin}
\title{abc}
\usage{
SldWin(Y, alpha, GLRSApprox, windowLength, maxCP, lowerLambda, upperLambda, breakPoint, 
	BetaFlg, W, Beta, crit, max_p, max_d, max_q, tol, maxIter, maxLev, innovDist, nsim1, nsim2, maxErr, minNonNAN, simType)
}
\arguments{
  \item{Y}{ abc }
  \item{alpha}{ abc }
  \item{GLRSApprox}{ abc }
  \item{windowLength}{ abc }
  \item{maxCP}{ abc }
  \item{lowerLambda}{ abc }
  \item{upperLambda}{ abc }
  \item{breakPoint}{ abc }
  \item{BetaFlg}{ abc }
  \item{W}{ abc }
  \item{Beta}{ abc }
  \item{crit}{ abc }
  \item{max_p}{ abc }
  \item{max_d}{ abc }
  \item{max_q}{ abc }
  \item{tol}{ abc }
  \item{maxIter}{ abc }
  \item{maxLev}{ abc }
  \item{innovDist}{ abc }
  \item{nsim1}{ abc }
  \item{nsim2}{ abc }
  \item{maxErr}{ abc }
  \item{minNonNAN}{ abc }
  \item{simType}{ abc }
}
\description{
	abc
}
\examples{
	\dontrun{
		# abc
	}
}